# backend/api/openapi.yaml
...
openapi: 3.0.1
info:
  title: WSO2 Motor Rally API
  version: 1.0.0
  description: |
    The rally game engine. Two identities share this API:

    - **Organizers** authenticate with an Asgardeo id token and set the rally
      up, then watch it run.
    - **Crews** authenticate with a team token minted by `POST /sessions/join`,
      the only unauthenticated write. There is no participant login: binding a
      vehicle is what authenticates a phone, and the one-active-phone rule
      means a second device is rejected with `409`.

    Every non-2xx response body is exactly `{"message": "..."}` — no `code`
    field. This mirrors the customer-portal contract the web app's
    `getApiErrorMessage()` already parses; the status code carries the machine-
    readable half.

    Paths carry no `/v1` segment: versioning is handled by Choreo API
    management, which publishes this definition under a versioned path.

    List endpoints are `POST /<resource>/search` with `{offset, limit,
    filters}`; the limit defaults to 20 and caps at 100.
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /events/{eventId}/leaderboard:
    parameters:
      - $ref: "#/components/parameters/EventId"
    get:
      tags: [Run views]
      summary: The pavilion leaderboard.
      description: |
        Highest score first; a tie goes to whoever finished earlier, and a team
        still running loses a tie against one that has finished. Teams with the
        same score and finish time share a rank.
      responses:
        "200":
          description: The ranked standings.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/LeaderboardEntry" }
  /events/{eventId}/monitor:
    parameters:
      - $ref: "#/components/parameters/EventId"
    get:
      tags: [Run views]
      summary: The live monitor snapshot.
      description: |
        Every vehicle in the event, whether or not a crew has bound to it, plus
        the open-alert count behind the dashboard's warning card.
      responses:
        "200":
          description: The snapshot.
          content:
            application/json:
              schema:
                type: object
                properties:
                  vehicles:
                    type: array
                    items: { $ref: "#/components/schemas/VehicleProgress" }
                  openAlerts: { type: integer }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        An Asgardeo id token for organizers, or the team token from
        `POST /sessions/join` for crews. The service tells them apart by issuer.
  parameters:
    EventId:
      name: eventId
      in: path
      required: true
      schema: { type: string, pattern: "^[0-9a-f]{32}$" }
...
  schemas:
...
    LeaderboardEntry:
      type: object
      properties:
        rank: { type: integer }
        vehicleCode: { type: string }
        teamName: { type: string }
        totalScore: { type: integer }
        finishTime: { type: string, format: date-time, nullable: true }
    VehicleProgress:
      type: object
      properties:
        vehicleCode: { type: string }
        teamName: { type: string }
        status: { type: string, enum: [ok, breakdown, device_issue] }
        sessionStatus:
          type: string
          description: Empty when no crew has bound to the vehicle.
          enum: ["", bound, active, finished]
        done: { type: integer }
        totalTasks: { type: integer }
        totalScore: { type: integer }
        lastLat: { type: number, nullable: true }
        lastLng: { type: number, nullable: true }
        lastSeenAt:
          type: string
          format: date-time
          nullable: true
          description: When the car's last accepted fix was taken.
        lastReportAt:
          type: string
          format: date-time
          nullable: true
          description: |
            When any phone in the car last reported, stamped on arrival. A
            replayed buffer carries old fix times, so this — not lastSeenAt —
            says whether the car is covered right now.
        coverageLost:
          type: boolean
          description: |
            A car on the course whose phones have all been silent for over 30
            seconds. Live changes arrive as coverage_lost / coverage_restored
            on the event topic.
...